Moved in-progress Primitive refactor to parallel module to ease navigation
